Add go server with registration and login endpoints.
Uses gin as a web server framework and gorm as an ORM, and PostgreSQL as
the database, just as with the previous server.

Authentication is done via JWT.

IDs are generated using snowflakes, and they will be distributed to
clients in base58 encoding, as this is a very short form encoding that
is easier to read than the 64-bit number (as was displayed before).