Add util to probe videos.
Currently probes MIME type and duration.

Duration is currently stored as an integer in seconds, therefore the
float value returned from ffprobe is rounded up - in the future, we may
wish to store the real value in the database so that we can choose to
round up or not when reading.

A map of ffmpeg formats to their MIME type is stored in the util - in
the future, this may be offloaded to a file, if we don't want to keep it
in the binary.

The map was sourced from a gist created in 2020:
https://gist.github.com/DusanBrejka/35238dccb5cefcc804de1c5a218ee004

The author writes that the command used to generate the map is as
follows (changed 'demuxer' -> 'muxer' to make it work for me):
```
ffmpeg -formats -hide_banner | 
  tail -n +5 |
  cut -c5- |
  cut -d' ' -f1 |
  xargs -i{} ffmpeg -hide_banner -h muxer={} |
  pcregrep -o2 -o4 -M '(Muxer (\w+) )|(Mime type:( .*).)'
```

We can choose to omit certain formats from the map to drop support for
them as desired. Currently, even images will upload successfully.

We can also filter images by duration or frame count - however, for the
time being, I see no harm in allowing images.