Reset all wrapped classes to one line.
Due to https://github.com/vuejs/language-tools/issues/3406, using
@vue/language-plugin-pug causes syntax errors when wrapping classes.

If this is fixed we could restore wrapping.

Alternatively, we could switch to HTML, however due to
https://github.com/prettier/prettier/issues/10918, prettier currently
does not wrap classes in HTML, so we would be back where we started.

The plan is to wait and see which technology evolves to be ergonomic
first, I guess :D.