Implement auto-generated anonymous users.
An anonymous user is used to track information - in particular,
currently an anonymous user is generated if a user uploads a video
whilst they are not logged in. As such, the video author can be kept
track of without requiring a registration.

If the user later registers or signs in, all of the data associated with
their anonymous user will be migrated to their user account. This is
known as an 'absorb' operation.

Anonymous users cannot be logged into. This is not explicitly enforced,
however anonymous users have a null hashed password, which cannot be
generated via bcrypt (as far as I know).

Without being able to log in, if a user ever loses the cookie associated
with the anonymous user (or the cookie expires), all of the associated
data will be lost without a server administrator interfering.

Therefore in the future we should add a warning banner for anonymous
users that have data to lose (such as uploaded videos), urging them to
create an account before their login token expires. We might also want
to increase token lifetime to 1 month or more (we could scope this to
anonymous users only).