Remove deleted_at field from database.
Currently we have no intention of marking data as deleted instead of
truly deleting it.

If we need soft delete in the future, we could bring it back, possibly
only for tables that require it.

Keeping the deleted_at field has the disadvantages that it slows down
writes as it is an index, and it also requires us to write more code to
delete data. It also slows down reads (although to what extent is
unknown) as they require additional conditions to get around
soft-deleted rows.