Implement view counter.
Views are counted when it is feasible that the user has watched a
significant portion of the video, currently configured at 60%.

To ensure this, when the user starts watching a new video, this is
recorded in the database. The number of bytes they have streamed so far
as well as the beginning of their watch session are both saved.

Once they have streamed 60% of the video file and 60% of the duration of
the video has passed since they first started, they are allowed to claim
a view.

Since browsers may sometimes stream an entire video in a single request,
it is not possible to verify the user's watch duration. Instead, a
client-side hint can be provided at .../still-watching via a POST
request, which will claim the view for the video, provided the
conditions are met.

An additional endpoint is available at .../required-watch-time-ms so
that the client can query when they should claim the view, which allows
the backend flexibility in changing the view requirements without
requiring a client update.

The view algorithm is intentionally conservative. It may / will not
count views under these circumstances:
 - A user is watching two or more (long) videos simultaneously.
   Interleaved streaming will result in the user's progress being
   constantly reset.
 - Multiple clients are logged in as the same user. The same situation
   as above will take place.
 - A user loops or streams the same video multiple times in a row
   (independent of time frame). Since view IDs have to be unique, and a
   view ID is generated when accessing a new video, multiple views on
   the same video can not be claimed in a row.
 - The user watches a short video in the browser, then watches another
   short video, and watches the first video again. Only one view will be
   counted for the first video. The browser will cache small videos,
   which means the 60% streaming requirement will not be met, as the
   video need not be downloaded again from the server.

A great advantage of this view algorithm is that, should a malicious
party attempt to forge a view, they will have to:
 1. Download 60% of a video.
 2. Wait 60% of the duration of the video.
 3. Make a request to claim a view.

Completing the steps above is no different to loading a video in the
browser and walking away from the computer until the view is
automatically counted, rendering the forgery process useless.