Add port forwarding from backend to frontend.
We will use this to debug HTML metadata injection shortly.

Devs can now choose to visit the client (port 5173) or the server (port
5000) when opening the site. Both will redirect all remaining requests
to each other. Visiting the client is recommended to avoid spam in the
backend request logs.

To prevent a redirect loop, 404 API requests are not forwarded to the
frontend (which can't answer them anyway).