Generate OpenGraph metadata for video pages.
We now have a video preview on social media websites.

To implement server-side metadata generation, a comment in the base HTML
marks where metadata should be inserted. To replace this comment, we
need to overwrite an existing HTML response, which is not supported by
gin by default.

Instead, we implemented a deferred response struct which records all
data written but stays open for modifications. At the end of the
rewriting middleware, the response struct's content is flushed to the
true HTTP response stream.