Imprint Streamfox logo on video previews.
To achieve this, we have a new .../preview endpoint specifically for
video previews in OpenGraph metadata. Using Go's image processing
library, we render the logo on top of the thumbnail.

We use a smaller version of the logo as thumbnails are low-resolution.

In the future, we may decide to pre-render previews on video upload and
keep them on disk, rather than re-generating them on every request. This
reduces server load and improves response speed, however it's less
flexible - any changes in the logo or preview format will only take
effect for new videos, unless an expensive and time-consuming operation
to re-generate all existing previews is executed.