Use a consistent error handling mechanism in controllers.
Introduced a new pattern of `if ok := checkError(params); !ok {`, which
makes code more concise and abstracts error handling, making the entire
system easier to read and maintain.

In addition, almost all errors were consolidated into predefined blocks,
so that all error messages can be maintained in one place.

We still have some work to do - in particular, validation errors and
errors that are being string-formatted are still not in the errors file.
We can consolidate those in the future, after applying some planned
changes.