Refactor error handling
Consolidate a consistent error response type - one containing generic
and specific errors. Specific errors are a map from each field in the
request body to a list of errors for it, whereas generic errors are a
list of errors loosely related to the request itself (including the
entire body).

Providing abstractions for this kind of generic error handling on the
frontend side converts verbose error handling code into just a few lines
of prose.