Stream enough video before claiming a view
For users with slow internet connections, a view might not be counted
because the client-side view timer fires before enough of the video has
been streamed.

In this change, on top of the existing timer, we also check 'progress'
events from the player to ensure that we have streamed the required
percentage of the video (as communicated by the server).

This is a 2nd pass on the view algorithm. Unfortunately there are still
a couple of issues that might require a 3rd pass:
 - Progress events from the player seem to malfunction in some browsers,
   particularly when the video has already been cached. We could resolve
   this by not relying on the browser events, instead polling the server
   until we have satisfied all view conditions.
 - The server checks bytes streamed, whereas the client checks duration
   streamed. If the byte distribution is not even, this might cause a
   discrepancy where the client has streamed enough of the video in
   duration but not yet in bytes, and the server will reject the view.
   This is not a huge issue as we prefer false negatives to false
   positives. It could be solved by checking duration instead of bytes
   on the server, or by finding a mechanism indicating how many bytes of
   the video have been downloaded on the client. Checking duration on
   the server would also mean implementing adaptive bitrate streaming is
   simpler (later on).
 - Close progress events can cause multiple view claims to be sent
   simultaneously. These are rejected by the server anyway, but finding
   a way to avoid them would be cleaner.