Add a prod docker environment
Run with `docker compose -f docker/prod/prodenv.yml up`.

Using docker compose, this spawns two containers, one for PostgreSQL,
and one for the server along with its static files.

Using multi-stage builds, we build the frontend and backend separately,
and only package the static files and final executable in the main
image, which produces a small 160MB production image (mostly taken up by
ffmpeg).

PostgreSQL is not included in the Streamfox image and therefore it
should always be spawned as a separate container or service, which
allows flexibility in where the database is deployed.