Change the port scheme
Until now, ports were mostly left as their defaults. However, we now
have 3 environments that a developer might be working in:
 - Local development environment (local-dev)
 - Docker development environment (docker-dev)
 - Docker production environment (docker-prod)

Depending on the needs, all 3 may need to be active at the same time. We
create a new port scheme to accomodate this.

In particular, we use a 4-number port scheme xyww.
 - x is always 8
 - y represents the environment: 6 (local-dev), 7 (docker-dev), or 8
   (docker-prod)
 - ww is a 2-digit number representing the particular application
   running, currently assigned as follows:
    - 01: backend
    - 02: frontend
    - 03: pgadmin

Note that postgres instances in docker containers are not exposed to the
host network, therefore they do not need to be identified by unique
ports. The default port 5432 has been retained.

To summarise, the ports that the developer will have exposed in the
given environments are:
 - local-dev:
   - 8601: backend
   - 8602: frontend
 - docker-dev:
   - 8701: backend
   - 8702: frontend
   - 8703: pgadmin
 - docker-prod:
   - 8801: backend

The developer can therefore run all 3 environments simultaneously, yet
test the applications independently.

Note that in the development environments, the developer can choose
whether to point their browser to the backend or frontend server, and
each will forward requests to the other as necessary.  Connecting
directly to the backend allows for testing server-side metadata
generation. However, connecting directly to the frontend should be
preferred as it will serve frontend resources faster.

Also note that there is space to add additional applications in the
future (up to a total of 99), such as a VS Code server in docker-dev.

The port scheme was designed to be straightforward whilst also using
ports that are unlikely to be occupied.