Implement a configuration-based file manager
All of the files managed by Streamfox are now encoded as a
human-readable file tree string in backend/files/base.go. When a file is
needed, it is referenced by a unique ID. The file tree structure can
therefore be easily modified without affecting any code that performs
operations on those files.