Set the video title to the name of the file on upload
By default, videos will have the name "Untitled Video". Previously we
set the video name when creating the video, however this would be
problematic when a video upload is retried multiple times with different
files (the video would keep the name of the first file). Now we set the
video name once upload is successful, so the video will always have the
name of the latest file used during upload.