Implement streamed uploads
Videos will upload in 100MB chunks to avoid keeping large video files
entirely in computer memory (which will usually crash the browser past a
certain point).

Chunked uploads work by sending repeated requests to the PUT
/videos/<id>/stream endpoint along with a Content-Range header
indicating which portion of the video is being uploaded.

Note that this endpoint is not completely robust; for example it is
possible for foreign clients to upload the video out-of-order.
Nevertheless, we are currently only concerned with the real web client.

In the future we could add configuration options in the user interface
which dictate the desired chunk size, allowing site admins to lower the
size if they're behind proxies which have a request size limit.