Use a helper to check for known content length
Not all requests know their content length ahead of time, which is why
we use a complex condition to check for validity before using the
content length.

Encapsulating the condition within a helper makes the code more
readable, and even helped to address a logic error (c.Request.Body !=
nil was wrong)!