Handle video URLs with an embedded timestamp
Allows linking to a specific timestamp by adding a 't' parameter to the
query, e.g. /watch/<id>?t=140 will begin playing the video from the 2
minutes and 20 seconds timestamp.

A player.load() call is needed as sometimes the source will not load
early enough, causing the time skip to be dropped.