Simplify the view algorithm
Reduce the view requirement to 1 minute and 30 seconds, instead of a
percentage of the video.

Longer videos will see less views counted on average with a percentage
algorithm. In addition, users may watch a video in more than one
sitting. Now we are more lenient towards counting these views (even
though they will be counted multiple times - indicating multiple
occasions on which the user has visited the video).

In addition, hint to the backend that we are watching every 30 seconds,
as well as at the start and end of a video.

Although initially continuous hinting seemed wasteful, emulating
progress tracking on the frontend is unreliable because some browsers
will sometimes miss progress events. We would also see console spam for
each progress event when the backend refused to accept a view.

Continuous hinting also significantly simplifies the code on both sides,
and allows the server complete discretion over the view counting
algorithm - only minor changes would be needed on the client to hint
more or less often.