Switch environment variables to envconfig package
envconfig allows the environment variables to be housed in a
strongly-typed struct, instead of being fetched individually all over
the place.

This makes the code clearer and more concise, and eliminates the need to
auto-generate enum strings.

I have also taken the creative liberty of moving the DB_NAME variable in
front of DB_USER, which matches the semantic order where the database is
accessed by name before being logged into.