Convert logged-in user badge to dropdown button
The dropdown houses a menu with actions that logged-in users can take
(currently, logging out only). Thus, the log out button no longer takes
up extra space in the navbar.

The dropdown button is extracted to a reusable component as an
alternative to the regular button. It neatly wraps the regular button
component with the dropdown component, forwarding all props to the
button.

Note that, while the generic dropdown makes the dropdown slot default
and names the button slot, the button dropdown does the opposite - the
button content slot is default and the dropdown template needs to be
specified explicitly.