Make video thumbnail anchors more semantic
Move the thumbnail link to the heading instead of the entire div. To
retain the entire thumbnail div area as clickable, add a child span to
the anchor which covers the div via `position: absolute` and `inset: 0`.

The cover workaround is extracted to well-named components so that the
pattern usage is clear, and as an added benefit, reusable in the future.

Moving the anchor to only link over the heading has semantic benefits -
for example, browsers will offer to copy the URL text, which will be
just the video heading, as opposed to all other text contained inside
(such as the duration, author, views, publish date, etc).

In addition, as per the HTML spec, anchor tags are not designed to have
block children, and while browsers are lenient, search engines will
lower the rank of the website for such violations.