Add user profile page
In the logged-in user dropdown, add a 'Profile' item which takes the
user to
their profile page. The profile page lists all of their uploaded videos,
along with a small icon indicating the visibility.

The profile page can be accessed for any user, however it will show
visibility only if you view your own page when you are logged in.
Viewing other users' pages will not show their private or unlisted
videos.

Currently there are no links to other users' profile pages, so to view
another user's page, you would have to uncover their ID and directly
visit the page by URL. In the future we will add a link to the profile
page whenever a user badge appears.