Auto-focus dynamically added tabs
When the user creates a new tab, they usually want to focus it by
default. However, we want to avoid always focusing the last added tab,
since this would cause the interface to initialize with the last tab
always selected, which is unlikely to be the correct tab for the user.

Hence, only focus tabs added after the component was mounted.