Add edit video option to user profile page
Alongside the existing video visibility icon, add a dropdown to all
videos on the user's personal profile page which includes a single
option - a button leading to the edit video page.

Since this is shown based on the same logic as the visibility icon,
rename the showVisibility prop to isPersonalPage as it is no longer
specific to the icon.

To accomodate the new menu, ensure the dropdown always fits its content,
and add a prop which allows left and right alignment (with left being
default).