Fetch dependencies as a separate step
Fetching dependencies before building the code ensures that small
changes to the code can be built without fetching dependencies all over
again.

Dependencies will only be re-fetched when they change or when the base
image changes.

Since the additional commands only apply to prerequisite stages, the
size of the final image will not be affected.