Rename compose environments to `compose.yml`
The 'devenv' and 'prodenv' files unnecessarily repeat the name of the
environment already specified by the directory they are placed in, which
makes them verbose and harder to work with when doing variable
substitution.

Removing the specifier leaves 'env' which could be ambiguous. Instead,
using the name 'compose' is widely accepted, even detected by editors
such as VS Code.