Implement user fetching
After a user is authenticated, allow fetching the current user at
/api/user.

All requests can receive the user authenticated for the request by
accepting a user parameter.

Also refactored the Authenticate error into EncodeJwt (to partner with
the new DecodeJwt error).

In addition, implemented Serialize / Deserialize for Id so that it
doesn't have to be manually converted into a string anywhere.