Reduce bcrypt cost
The default cost of 12 takes too long to execute. Reduce to 10, the same
as the default cost in Go.

With a cost of 12, login takes anywhere from 200 - 500ms to execute on a
Ryzen 7 3800X in release, and up to 4s in debug. In comparison, a cost
of 10 takes 50ms in release and 200ms in debug.