Create a filesystem generator procedural macro
Given a filesystem layout in a custom syntax (as a string), the new
procedural macro will generate a struct impl that allows each distinct
file declared to be resolved by setting values for the relevant
placeholders in the file path.

Moving forward, the resolved file struct can be extended to support the
desired operations.

A declarative filesystem layout is useful because it allows structural
re-organisation with little effort.