Sign JWTs with the user's password hash
Instead of hardcoding the JWT signing secret in the binary, use each
user's password hash as the secret. When the user changes password, the
secret will automatically change, thus automatically revoking all
existing logins.

In addition, these hashes are less guessable (than using a hardcoded
string or even a single secret for the entire app), since they are
cryptographically secure random strings and different for every user.

To validate the token, we now need to fetch the user from the ID stored
in the token itself. To make this easier, store the user ID in the token
header under the key "kid".