Convert User extractor into middleware
Extracting the user in a middleware and passing it to handlers via
extensions is the correct approach for this job, as we want to avoid
fetching the user every time it is required by a middleware in the
handler chain.

Also split requests into authenticated (which run the user extension)
and unauthenticated (which do not need the user and avoid running the
extension at all).