Fix issues with fetching all videos
The following issues were overlooked:
 - views was inner joined when it should have been left joined;
 - all integers sourced from the database should be signed and the HTTP
   response should be signed, with a conversion;
 - only complete and public videos should be shown (via a filter); and
 - a DateTimeWithTimeZone should be sourced from the database and
   appropriately converted.

In addition we source the likes and dislikes from the databse, filled in
as zeroes until we implement the likes and dislikes tables.

Note that SeaORM has a bug that requires all 'as' calls to be placed
before find_also_related, otherwise we must append A_ to every alias.