backend: Allow passing the path to the config file
Currently, configuration values are always loaded from a `.env` file in
the current working directory. Instead, allow specifying the path to the
configuration file as a command-line argument.

Note that this is an unstructured way of using command-line arguments;
in the case that we add more arguments, we should properly parse them
into a structure for clarity and ease of use.